refactor: move fiat-shamir related methods into `verifier.rs` and rename `cryptography.rs` to `kzg_open.rs`
