+++ crates/eip4844/src/lib.rs
@@ -10,14 +10,14 @@ mod prover;
 pub(crate) mod verifier;
 
 pub mod constants;
-mod cryptography;
+mod kzg_open;
 mod serialization;
 mod trusted_setup;
 
 pub use errors::{Error, SerializationError, VerifierError};
 pub use trusted_setup::TrustedSetup;
 
-use crate::cryptography::{prover::Prover, verifier::Verifier};
+use crate::kzg_open::{prover::Prover, verifier::Verifier};
 
 /// BlobRef denotes a references to an opaque Blob.
 ///
