+++ crates/eip4844/src/verifier.rs
@@ -1,8 +1,13 @@
+use std::iter::successors;
+
+use bls12_381::{
+    ff::{Field, PrimeField},
+    reduce_bytes_to_scalar_bias, G1Point, Scalar,
+};
+use sha2::{Digest, Sha256};
+
 use crate::{
-    cryptography::{
-        compute_fiat_shamir_challenge,
-        verifier::{compute_evaluation, compute_r_powers_for_verify_kzg_proof_batch},
-    },
+    kzg_open::verifier::compute_evaluation,
     serialization::{
         deserialize_blob_to_scalars, deserialize_bytes_to_scalar, deserialize_compressed_g1,
     },
@@ -134,4 +139,120 @@ impl Context {
 
         Ok(())
     }
+}
+
+/// Compute Fiat-Shamir challenge of a blob KZG proof.
+///
+/// The matching function in the specs is: https://github.com/ethereum/consensus-specs/blob/017a8495f7671f5fff2075a9bfc9238c1a0982f8/specs/deneb/polynomial-commitments.md#compute_challenge
+pub(crate) fn compute_fiat_shamir_challenge(blob: BlobRef, commitment: KZGCommitment) -> Scalar {
+    // DomSepProtocol is a Domain Separator to identify the protocol.
+    //
+    // It matches [FIAT_SHAMIR_PROTOCOL_DOMAIN] in the spec.
+    //
+    // [FIAT_SHAMIR_PROTOCOL_DOMAIN]: https://github.com/ethereum/consensus-specs/blob/017a8495f7671f5fff2075a9bfc9238c1a0982f8/specs/deneb/polynomial-commitments.md#blob
+    const DOMAIN_SEP: &str = "FSBLOBVERIFY_V1_";
+
+    let bytes_per_commitment = G1Point::compressed_size();
+    let bytes_per_blob = blob.len();
+
+    let bytes_per_field_element = Scalar::NUM_BITS.div_ceil(8) as usize;
+    let field_elements_per_blob = blob.len() / bytes_per_field_element;
+
+    let hash_input_size = DOMAIN_SEP.len()
+            + 2 * size_of::<u64>() // polynomial bound
+            + bytes_per_blob // blob
+            + bytes_per_commitment // commitment
+            ;
+
+    let mut hash_input: Vec<u8> = Vec::with_capacity(hash_input_size);
+
+    hash_input.extend(DOMAIN_SEP.as_bytes());
+    hash_input.extend(0u64.to_be_bytes());
+    hash_input.extend((field_elements_per_blob as u64).to_be_bytes());
+    hash_input.extend(blob);
+    hash_input.extend(commitment);
+
+    assert_eq!(hash_input.len(), hash_input_size);
+    let mut hasher = Sha256::new();
+    hasher.update(hash_input);
+    let result: [u8; 32] = hasher.finalize().into();
+
+    // For randomization, we only need a 128 bit scalar, since this is used for batch verification.
+    // See for example, the randomizers section in : https://cr.yp.to/badbatch/badbatch-20120919.pdf
+    //
+    // This is noted because when we convert a 256 bit hash to a scalar, a bias will be introduced.
+    // This however does not affect our security guarantees because the bias is negligible given we
+    // want a uniformly random 128 bit integer.
+    //
+    // Also there is a negligible probably that the scalar is zero, so we do not handle this case here.
+    reduce_bytes_to_scalar_bias(result)
+}
+
+/// Compute random linear combination challenge scalars for batch verification.
+///
+/// The matching function in the specs is: https://github.com/ethereum/consensus-specs/blob/017a8495f7671f5fff2075a9bfc9238c1a0982f8/specs/deneb/polynomial-commitments.md#verify_kzg_proof_batch
+pub fn compute_r_powers_for_verify_kzg_proof_batch(
+    domain_size: usize,
+    commitments: &[KZGCommitment],
+    zs: &[Scalar],
+    ys: &[Scalar],
+    proofs: &[KZGProof],
+) -> Vec<Scalar> {
+    // DomSepProtocol is a Domain Separator to identify the protocol.
+    //
+    // It matches [RANDOM_CHALLENGE_KZG_BATCH_DOMAIN] in the spec.
+    //
+    // [RANDOM_CHALLENGE_KZG_BATCH_DOMAIN]: https://github.com/ethereum/consensus-specs/blob/017a8495f7671f5fff2075a9bfc9238c1a0982f8/specs/deneb/polynomial-commitments.md#blob
+    const DOMAIN_SEP: &str = "RCKZGBATCH___V1_";
+
+    let bytes_per_commitment = G1Point::compressed_size();
+    let bytes_per_field_element = Scalar::NUM_BITS.div_ceil(8) as usize;
+
+    let n = commitments.len();
+
+    let hash_input_size = DOMAIN_SEP.len()
+        + size_of::<u64>() // polynomial bound
+        + size_of::<u64>() // batch size
+        + n * (
+            bytes_per_commitment // commitment
+            + bytes_per_field_element // z
+            + bytes_per_field_element // y
+            + bytes_per_commitment // proof
+        );
+
+    let mut hash_input: Vec<u8> = Vec::with_capacity(hash_input_size);
+
+    hash_input.extend(DOMAIN_SEP.as_bytes());
+    hash_input.extend((domain_size as u64).to_be_bytes());
+    hash_input.extend((n as u64).to_be_bytes());
+    commitments
+        .iter()
+        .zip(zs)
+        .zip(ys)
+        .zip(proofs)
+        .for_each(|(((commitment, z), y), proof)| {
+            hash_input.extend(commitment);
+            hash_input.extend(z.to_bytes_be());
+            hash_input.extend(y.to_bytes_be());
+            hash_input.extend(proof);
+        });
+
+    assert_eq!(hash_input.len(), hash_input_size);
+    let mut hasher = Sha256::new();
+    hasher.update(hash_input);
+    let result: [u8; 32] = hasher.finalize().into();
+
+    // For randomization, we only need a 128 bit scalar, since this is used for batch verification.
+    // See for example, the randomizers section in : https://cr.yp.to/badbatch/badbatch-20120919.pdf
+    //
+    // This is noted because when we convert a 256 bit hash to a scalar, a bias will be introduced.
+    // This however does not affect our security guarantees because the bias is negligible given we
+    // want a uniformly random 128 bit integer.
+    //
+    // Also there is a negligible probably that the scalar is zero, so we do not handle this case here.
+    let r = reduce_bytes_to_scalar_bias(result);
+
+    successors(Some(Scalar::ONE), |power| Some(*power * r))
+        .take(n)
+        .collect()
 }
